+++ tgcli/request/bot.py
@@ -80,6 +80,7 @@ class SendPollRequest(BotRequest):
         options: typing.Iterable[str],
         allows_multiple_answers: bool,
         is_anonymous: bool,
+        open_period: int,
         disable_notification: bool = False,
     ):
         try:
@@ -95,6 +96,7 @@ class SendPollRequest(BotRequest):
             "options": tuple(options),
             "allows_multiple_answers": bool(allows_multiple_answers),
             "is_anonymous": bool(is_anonymous),
+            **({"open_period": int(open_period) if open_period else {}}),
             "disable_notification": bool(disable_notification),
         }
         self.prepare_body(data=None, files=None, json=payload)